backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
redis
passlib[bcrypt]
python-jose[cryptography]
python-multipart
PyJWT
